app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.3.0"
description = ""
authors = ["Alexander Demure"]

[tool.poetry.dependencies]

python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.5"
structlog = "^20.1.0"
python-multipart = "^0.0.5"
tenacity = "^6.2.0"
pydantic = "^1.4"
python-jose = "^3.1.0"
passlib = "^1.7.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
